requirements.txt
flask==3.1.0
openai==1.68.0
python-dotenv==1.0.1
langgraph>=0.2.0
langchain-core>=0.3.0
duckduckgo-search>=6.0.0
# Google Tasks API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
